{"schema":"decodex/commit/1","summary":"Split worktree and program modules","authority":"manual"}
