+++ apps/decodex/src/orchestrator/program_reconciler/model.rs
@@ -8,27 +8,25 @@ use crate::{
 	},
 };
 
-pub(in crate::orchestrator) const PROGRAM_DISPATCH_SELECTED_SCHEMA: &str =
-	"decodex.program_dispatch_selected/1";
-pub(in crate::orchestrator) const PROGRAM_DISPATCH_SELECTED_EVENT_TYPE: &str =
-	"program_dispatch_selected";
+pub(crate) const PROGRAM_DISPATCH_SELECTED_SCHEMA: &str = "decodex.program_dispatch_selected/1";
+pub(crate) const PROGRAM_DISPATCH_SELECTED_EVENT_TYPE: &str = "program_dispatch_selected";
 
 #[derive(Clone)]
-pub(in crate::orchestrator) struct RefreshedExecutionProgram {
-	pub(in crate::orchestrator) record: ExecutionProgramRecord,
-	pub(in crate::orchestrator) program: ExecutionProgram,
-	pub(in crate::orchestrator) issues_by_node: BTreeMap<String, ProgramIssueSnapshot>,
+pub(crate) struct RefreshedExecutionProgram {
+	pub(crate) record: ExecutionProgramRecord,
+	pub(crate) program: ExecutionProgram,
+	pub(crate) issues_by_node: BTreeMap<String, ProgramIssueSnapshot>,
 }
 
 #[derive(Clone)]
-pub(in crate::orchestrator) struct ProgramIssueSnapshot {
-	pub(in crate::orchestrator) issue: TrackerIssue,
-	pub(in crate::orchestrator) has_active_label: bool,
-	pub(in crate::orchestrator) has_opt_out_label: bool,
-	pub(in crate::orchestrator) has_needs_attention_label: bool,
-	pub(in crate::orchestrator) has_open_tracker_blockers: bool,
-	pub(in crate::orchestrator) has_generic_dispatch_briefing: bool,
-	pub(in crate::orchestrator) has_post_review_lifecycle: bool,
+pub(crate) struct ProgramIssueSnapshot {
+	pub(crate) issue: TrackerIssue,
+	pub(crate) has_active_label: bool,
+	pub(crate) has_opt_out_label: bool,
+	pub(crate) has_needs_attention_label: bool,
+	pub(crate) has_open_tracker_blockers: bool,
+	pub(crate) has_generic_dispatch_briefing: bool,
+	pub(crate) has_post_review_lifecycle: bool,
 }
 impl ProgramIssueSnapshot {
 	pub(super) fn linear_mapping(&self) -> Result<ExecutionLinearIssueMapping> {
@@ -46,25 +44,25 @@ impl ProgramIssueSnapshot {
 	}
 }
 
-pub(in crate::orchestrator) struct ProgramIssueSnapshotInput<'a, T>
+pub(crate) struct ProgramIssueSnapshotInput<'a, T>
 where
 	T: IssueTracker + ?Sized,
 {
-	pub(in crate::orchestrator) tracker: &'a T,
-	pub(in crate::orchestrator) state_store: &'a StateStore,
-	pub(in crate::orchestrator) service_id: &'a str,
-	pub(in crate::orchestrator) workflow: &'a WorkflowDocument,
-	pub(in crate::orchestrator) issue: &'a TrackerIssue,
+	pub(crate) tracker: &'a T,
+	pub(crate) state_store: &'a StateStore,
+	pub(crate) service_id: &'a str,
+	pub(crate) workflow: &'a WorkflowDocument,
+	pub(crate) issue: &'a TrackerIssue,
 }
 
 #[derive(Default)]
-pub(in crate::orchestrator) struct ProgramSchedulerSummary {
-	pub(in crate::orchestrator) programs_evaluated: usize,
-	pub(in crate::orchestrator) programs_updated: usize,
-	pub(in crate::orchestrator) dispatchable_nodes: usize,
+pub(crate) struct ProgramSchedulerSummary {
+	pub(crate) programs_evaluated: usize,
+	pub(crate) programs_updated: usize,
+	pub(crate) dispatchable_nodes: usize,
 }
 
-pub(in crate::orchestrator) struct ProgramSchedulerSelection {
-	pub(in crate::orchestrator) selected: Option<SelectedIssueRunCandidate>,
-	pub(in crate::orchestrator) summary: ProgramSchedulerSummary,
+pub(crate) struct ProgramSchedulerSelection {
+	pub(crate) selected: Option<SelectedIssueRunCandidate>,
+	pub(crate) summary: ProgramSchedulerSummary,
 }
