+++ apps/decodex/src/state/tests/persistent_events.rs
@@ -1,818 +1,4 @@
-use std::{
-	sync::{Arc, Barrier},
-	thread,
-};
-
-use rusqlite::{Connection, Result};
-use tempfile::TempDir;
-
-use crate::{
-	state::{
-		ProjectRegistration, StateStore,
-		tests::{self, IN_PROGRESS_STATE},
-	},
-	tracker::records::{LinearExecutionEventIdentity, LinearExecutionEventRecord},
-};
-
-#[test]
-fn persistent_event_appenders_can_write_distinct_runs_concurrently() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let first = StateStore::open(&state_path).expect("first state store should open");
-	let second = StateStore::open(&state_path).expect("second state store should open");
-
-	first.record_run_attempt("run-a", "PUB-101", 1, "running").expect("first run should record");
-	second.record_run_attempt("run-b", "PUB-102", 1, "running").expect("second run should record");
-
-	let barrier = Arc::new(Barrier::new(2));
-	let first_barrier = Arc::clone(&barrier);
-	let first_writer = thread::spawn(move || {
-		first_barrier.wait();
-
-		for sequence_number in 1..=40 {
-			first
-				.append_event("run-a", sequence_number, "item/agentMessage/delta", "{}")
-				.expect("first event writer should append");
-		}
-	});
-	let second_writer = thread::spawn(move || {
-		barrier.wait();
-
-		for sequence_number in 1..=40 {
-			second
-				.append_event("run-b", sequence_number, "item/agentMessage/delta", "{}")
-				.expect("second event writer should append");
-		}
-	});
-
-	first_writer.join().expect("first event writer should finish");
-	second_writer.join().expect("second event writer should finish");
-
-	let reopened = StateStore::open(&state_path).expect("reopened state store should open");
-
-	assert_eq!(reopened.event_count("run-a").expect("first event count should load"), 40);
-	assert_eq!(reopened.event_count("run-b").expect("second event count should load"), 40);
-}
-
-#[test]
-fn persistent_append_event_does_not_refresh_full_event_journal() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let first = StateStore::open(&state_path).expect("first state store should open");
-	let second = StateStore::open(&state_path).expect("second state store should open");
-
-	first.record_run_attempt("run-a", "PUB-101", 1, "running").expect("first run should record");
-	second.record_run_attempt("run-b", "PUB-102", 1, "running").expect("second run should record");
-	second
-		.append_event("run-b", 1, "item/agentMessage/delta", "{}")
-		.expect("second store should append an unrelated event");
-	first
-		.append_event("run-a", 1, "item/agentMessage/delta", "{}")
-		.expect("first store should append without full journal refresh");
-
-	let state = first.inner.lock().expect("test should inspect the local cache");
-
-	assert!(
-		!state.events.contains_key("run-b"),
-		"append_event should not refresh the full persistent event journal into the local cache"
-	);
-
-	drop(state);
-
-	let reopened = StateStore::open(&state_path).expect("reopened state store should open");
-
-	assert_eq!(reopened.event_count("run-a").expect("first event count should load"), 1);
-	assert_eq!(reopened.event_count("run-b").expect("second event count should load"), 1);
-}
-
-#[test]
-fn persistent_open_backfills_protocol_event_summaries_from_legacy_journal() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-
-	{
-		let store = StateStore::open(&state_path).expect("state store should create schema");
-
-		store.record_run_attempt("run-legacy", "PUB-101", 1, "running").expect("run should record");
-	}
-
-	let connection = Connection::open(&state_path).expect("sqlite should open");
-
-	connection
-		.execute("DELETE FROM schema_meta WHERE key = 'schema_version'", [])
-		.expect("schema version should reset");
-	connection
-		.execute(
-			"DELETE FROM schema_meta
-			 WHERE key = 'migration:protocol_event_summaries_from_events:v12'",
-			[],
-		)
-		.expect("protocol summary migration marker should reset");
-	connection
-		.execute("DELETE FROM protocol_event_summaries", [])
-		.expect("summary rows should clear");
-	connection
-		.execute(
-			"INSERT INTO protocol_events (
-					run_id, sequence_number, event_type, payload_sha256, created_at, created_at_unix
-				) VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-			rusqlite::params![
-				"run-legacy",
-				1_i64,
-				"turn/started",
-				"sha-1",
-				"2026-06-17T00:00:00Z",
-				1_i64,
-			],
-		)
-		.expect("legacy event should insert");
-	connection
-		.execute(
-			"INSERT INTO protocol_events (
-					run_id, sequence_number, event_type, payload_sha256, created_at, created_at_unix
-				) VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-			rusqlite::params![
-				"run-legacy",
-				2_i64,
-				"turn/completed",
-				"sha-2",
-				"2026-06-17T00:00:01Z",
-				2_i64,
-			],
-		)
-		.expect("legacy event should insert");
-
-	let reopened = StateStore::open(&state_path).expect("state store should migrate");
-
-	assert_eq!(reopened.event_count("run-legacy").expect("event count should load"), 2);
-
-	let summary = connection
-		.query_row(
-			"SELECT event_count, last_sequence_number, last_event_type
-			 FROM protocol_event_summaries
-			 WHERE run_id = 'run-legacy'",
-			[],
-			|row| Ok((row.get::<_, i64>(0)?, row.get::<_, i64>(1)?, row.get::<_, String>(2)?)),
-		)
-		.expect("summary should persist");
-
-	assert_eq!(summary, (2, 2, String::from("turn/completed")));
-}
-
-#[test]
-fn persistent_open_keeps_protocol_backfill_marker_when_later_migration_fails() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-
-	{
-		let store = StateStore::open(&state_path).expect("state store should create schema");
-
-		store.record_run_attempt("run-legacy", "PUB-102", 1, "running").expect("run should record");
-	}
-
-	let connection = Connection::open(&state_path).expect("sqlite should open");
-	let mut removed_field_payload = serde_json::to_value(tests::latent_decision_contract_fixture())
-		.expect("fixture should encode as JSON");
-
-	removed_field_payload["contract_id"] = serde_json::json!("removed-field-invalid-contract");
-	removed_field_payload["status"] = serde_json::json!("accepted_promoted");
-	removed_field_payload["execution_readiness"]["ready_for_issue_shaping"] =
-		serde_json::json!(false);
-
-	let readiness = removed_field_payload
-		.get_mut("execution_readiness")
-		.expect("readiness should exist")
-		.as_object_mut()
-		.expect("readiness should be an object");
-
-	readiness.remove("proposed_issues");
-	readiness.insert(
-		String::from("proposed_issue_summaries"),
-		serde_json::json!(["Removed-field invalid summary."]),
-	);
-	connection
-		.execute("UPDATE schema_meta SET value = '11' WHERE key = 'schema_version'", [])
-		.expect("schema version should mark removed-field state");
-	connection
-		.execute(
-			"DELETE FROM schema_meta
-			 WHERE key = 'migration:protocol_event_summaries_from_events:v12'",
-			[],
-		)
-		.expect("protocol summary migration marker should reset");
-	connection
-		.execute("DELETE FROM protocol_event_summaries", [])
-		.expect("summary rows should clear");
-	connection
-		.execute(
-			"INSERT INTO protocol_events (
-					run_id, sequence_number, event_type, payload_sha256, created_at, created_at_unix
-				) VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
-			rusqlite::params![
-				"run-legacy",
-				1_i64,
-				"turn/started",
-				"sha-1",
-				"2026-06-17T00:00:00Z",
-				1_i64,
-			],
-		)
-		.expect("legacy event should insert");
-	connection
-		.execute(
-			"INSERT INTO decision_contracts (
-					project_id, contract_id, source_issue_id, status, payload_json, created_at,
-					created_at_unix, updated_at, updated_at_unix
-				) VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9)",
-			rusqlite::params![
-				"decodex",
-				"removed-field-invalid-contract",
-				"XY-BAD",
-				"accepted_promoted",
-				serde_json::to_string(&removed_field_payload)
-					.expect("removed-field payload should serialize"),
-				"2026-06-17T00:00:00Z",
-				1_i64,
-				"2026-06-17T00:00:00Z",
-				1_i64,
-			],
-		)
-		.expect("invalid removed-field decision contract row should insert");
-
-	assert!(
-		StateStore::open(&state_path).is_err(),
-		"invalid removed-field decision contract migration should still fail closed"
-	);
-
-	let marker: String = connection
-		.query_row(
-			"SELECT value FROM schema_meta
-			 WHERE key = 'migration:protocol_event_summaries_from_events:v12'",
-			[],
-			|row| row.get(0),
-		)
-		.expect("protocol backfill marker should persist after later migration failure");
-	let event_count: i64 = connection
-		.query_row(
-			"SELECT event_count FROM protocol_event_summaries WHERE run_id = 'run-legacy'",
-			[],
-			|row| row.get(0),
-		)
-		.expect("protocol summary should persist after later migration failure");
-
-	assert_eq!(marker, "completed");
-	assert_eq!(event_count, 1);
-}
-
-#[test]
-fn persistent_open_preserves_future_schema_version() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-
-	StateStore::open(&state_path).expect("state store should create schema");
-
-	let connection = Connection::open(&state_path).expect("sqlite should open");
-
-	connection
-		.execute("UPDATE schema_meta SET value = '13' WHERE key = 'schema_version'", [])
-		.expect("future schema version should set");
-
-	StateStore::open(&state_path).expect("state store should reopen");
-
-	let version: String = connection
-		.query_row("SELECT value FROM schema_meta WHERE key = 'schema_version'", [], |row| {
-			row.get(0)
-		})
-		.expect("schema version should read");
-
-	assert_eq!(version, "13");
-}
-
-#[test]
-fn protocol_event_replay_is_idempotent_when_payload_matches() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let store = StateStore::open(&state_path).expect("state store should open");
-	let payload = r#"{"threadId":"thread-1","attemptNumber":4}"#;
-
-	store
-		.append_event("run-1", 32, "thread/archive", payload)
-		.expect("first archive event should append");
-	store
-		.append_event("run-1", 32, "thread/archive", payload)
-		.expect("matching archive replay should be idempotent");
-
-	assert_eq!(store.event_count("run-1").expect("event count should load"), 1);
-
-	let connection = Connection::open(&state_path).expect("sqlite should open");
-	let payload_sha256: String = connection
-		.query_row(
-			"SELECT payload_sha256 FROM protocol_events WHERE run_id = 'run-1' AND sequence_number = 32",
-			[],
-			|row| row.get(0),
-		)
-		.expect("payload digest should read");
-
-	assert_eq!(payload_sha256.len(), 64);
-	assert!(payload_sha256.chars().all(|character| character.is_ascii_hexdigit()));
-}
-
-#[test]
-fn sparse_protocol_event_replay_keeps_leased_run_summary_count_exact() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let store = StateStore::open(&state_path).expect("state store should open");
-	let payload = r#"{"threadId":"thread-1","attemptNumber":4}"#;
-
-	store.record_run_attempt("run-1", "PUB-101", 1, "running").expect("run attempt should persist");
-	store
-		.upsert_lease("pubfi", "PUB-101", "run-1", IN_PROGRESS_STATE)
-		.expect("lease should persist");
-	store
-		.append_event("run-1", 32, "thread/archive", payload)
-		.expect("first sparse event should append");
-	store
-		.append_event("run-1", 32, "thread/archive", payload)
-		.expect("matching sparse replay should be idempotent");
-
-	let runs = store.list_leased_runs("pubfi").expect("leased runs should load");
-
-	assert_eq!(runs.len(), 1);
-	assert_eq!(runs[0].run_id(), "run-1");
-	assert_eq!(runs[0].event_count(), 1);
-	assert_eq!(runs[0].last_event_type(), Some("thread/archive"));
-}
-
-#[test]
-fn protocol_event_sequence_conflict_rejects_changed_payload() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let store = StateStore::open(&state_path).expect("state store should open");
-
-	store
-		.append_event("run-1", 32, "thread/archive", r#"{"threadId":"thread-1"}"#)
-		.expect("first archive event should append");
-
-	let error = store
-		.append_event("run-1", 32, "thread/archive", r#"{"threadId":"thread-2"}"#)
-		.expect_err("changed payload at the same sequence should be rejected");
-
-	assert!(
-		error.to_string().contains("conflicts with an existing runtime journal event"),
-		"unexpected error: {error:?}",
-	);
-	assert_eq!(store.event_count("run-1").expect("event count should load"), 1);
-}
-
-#[test]
-fn late_protocol_events_after_thread_archive_use_discarded_sequence_namespace() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let store = StateStore::open(&state_path).expect("state store should open");
-
-	store.record_run_attempt("run-1", "PUB-101", 1, "failed").expect("run attempt should record");
-	store
-		.upsert_worktree("pubfi", "PUB-101", "x/pubfi-pub-101", "/tmp/worktrees/pub-101")
-		.expect("worktree should record");
-	store
-		.append_event("run-1", 32, "thread/archive", r#"{"threadId":"thread-1"}"#)
-		.expect("archive event should append");
-	store
-		.append_event("run-1", 32, "item/started", r#"{"itemId":"item-1"}"#)
-		.expect("late item start should be discarded without conflict");
-	store
-		.append_event("run-1", 33, "item/completed", r#"{"itemId":"item-1"}"#)
-		.expect("later item completion should also be discarded");
-
-	let runs = store.list_recent_runs("pubfi", 10).expect("recent runs should load");
-
-	assert_eq!(runs.len(), 1);
-	assert_eq!(runs[0].event_count(), 3);
-	assert_eq!(runs[0].last_event_type(), Some("thread/archive"));
-	assert_eq!(store.event_count("run-1").expect("event count should load"), 3);
-
-	let connection = Connection::open(&state_path).expect("sqlite should open");
-	let mut statement = connection
-		.prepare(
-			"SELECT sequence_number, event_type FROM protocol_events \
-			 WHERE run_id = 'run-1' ORDER BY sequence_number",
-		)
-		.expect("protocol rows should prepare");
-	let rows = statement
-		.query_map([], |row| Ok((row.get::<_, i64>(0)?, row.get::<_, String>(1)?)))
-		.expect("protocol rows should query")
-		.collect::<Result<Vec<_>>>()
-		.expect("protocol rows should collect");
-	let discarded_rows = rows
-		.iter()
-		.filter(|(_sequence, event_type)| {
-			event_type.as_str() == "protocol/post_archive_event/discarded"
-		})
-		.collect::<Vec<_>>();
-
-	assert!(rows.iter().any(|row| row == &(32, String::from("thread/archive"))));
-	assert_eq!(discarded_rows.len(), 2);
-	assert!(discarded_rows.iter().all(|(sequence, _event_type)| *sequence < 0));
-	assert!(!rows.iter().any(|(_sequence, event_type)| event_type == "item/started"));
-	assert!(!rows.iter().any(|(_sequence, event_type)| event_type == "item/completed"));
-}
-
-#[test]
-fn post_archive_discard_append_rebuilds_stale_compacted_protocol_summary() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let store = StateStore::open(&state_path).expect("state store should open");
-
-	store.record_run_attempt("run-1", "PUB-101", 1, "failed").expect("run attempt should record");
-	store
-		.upsert_worktree("pubfi", "PUB-101", "x/pubfi-pub-101", "/tmp/worktrees/pub-101")
-		.expect("worktree should record");
-	store
-		.append_event("run-1", 32, "thread/archive", r#"{"threadId":"thread-1"}"#)
-		.expect("archive event should append");
-
-	let connection = Connection::open(&state_path).expect("sqlite should open");
-
-	connection
-		.execute(
-			"UPDATE protocol_event_summaries
-			 SET event_count = 1,
-			     last_sequence_number = 32,
-			     last_event_type = 'thread/archive'
-			 WHERE run_id = 'run-1'",
-			[],
-		)
-		.expect("compacted summary should be pinned stale");
-	store
-		.append_event("run-1", 32, "item/started", r#"{"itemId":"item-1"}"#)
-		.expect("late item start should be discarded without conflict");
-
-	assert_eq!(store.event_count("run-1").expect("event count should load"), 2);
-
-	let compacted_event_count: i64 = connection
-		.query_row(
-			"SELECT event_count FROM protocol_event_summaries WHERE run_id = 'run-1'",
-			[],
-			|row| row.get(0),
-		)
-		.expect("compacted summary should load");
-
-	assert_eq!(compacted_event_count, 2);
-}
-
-#[test]
-fn pre_archive_protocol_replays_remain_idempotent_after_archive() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let store = StateStore::open(&state_path).expect("state store should open");
-
-	store.record_run_attempt("run-1", "PUB-101", 1, "failed").expect("run attempt should record");
-	store
-		.upsert_worktree("pubfi", "PUB-101", "x/pubfi-pub-101", "/tmp/worktrees/pub-101")
-		.expect("worktree should record");
-	store
-		.append_event("run-1", 31, "item/started", r#"{"itemId":"item-1"}"#)
-		.expect("pre-archive item start should append");
-	store
-		.append_event("run-1", 32, "thread/archive", r#"{"threadId":"thread-1"}"#)
-		.expect("archive event should append");
-	store
-		.append_event("run-1", 31, "item/started", r#"{"itemId":"item-1"}"#)
-		.expect("pre-archive replay should stay idempotent after archive");
-
-	let runs = store.list_recent_runs("pubfi", 10).expect("recent runs should load");
-
-	assert_eq!(runs.len(), 1);
-	assert_eq!(runs[0].event_count(), 2);
-	assert_eq!(runs[0].last_event_type(), Some("thread/archive"));
-	assert_eq!(store.event_count("run-1").expect("event count should load"), 2);
-}
-
-#[test]
-fn late_protocol_events_after_archive_discard_are_restart_safe() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let store = StateStore::open(&state_path).expect("state store should open");
-
-	store.record_run_attempt("run-1", "PUB-101", 1, "failed").expect("run attempt should record");
-	store
-		.upsert_worktree("pubfi", "PUB-101", "x/pubfi-pub-101", "/tmp/worktrees/pub-101")
-		.expect("worktree should record");
-	store
-		.append_event("run-1", 7, "thread/archive/discarded", r#"{"threadId":"thread-1"}"#)
-		.expect("discarded archive event should append");
-
-	drop(store);
-
-	let restarted = StateStore::open_lazy(&state_path).expect("lazy store should open");
-
-	restarted
-		.append_event("run-1", 8, "item/commandExecution/outputDelta", r#"{"delta":"late output"}"#)
-		.expect("late output after restart should be discarded without conflict");
-
-	let runs = restarted.list_recent_runs("pubfi", 10).expect("recent runs should load");
-
-	assert_eq!(runs.len(), 1);
-	assert_eq!(runs[0].event_count(), 2);
-	assert_eq!(runs[0].last_event_type(), Some("thread/archive/discarded"));
-	assert_eq!(restarted.event_count("run-1").expect("event count should load"), 2);
-}
-
-#[test]
-fn persistent_run_attempt_update_does_not_refresh_full_event_journal() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let first = StateStore::open(&state_path).expect("first state store should open");
-	let second = StateStore::open(&state_path).expect("second state store should open");
-
-	second.record_run_attempt("run-b", "PUB-102", 1, "running").expect("second run should record");
-	second
-		.append_event("run-b", 1, "item/agentMessage/delta", "{}")
-		.expect("second store should append an unrelated event");
-	first.record_run_attempt("run-a", "PUB-101", 1, "running").expect("first run should record");
-	first.update_run_thread("run-a", "thread-a").expect("first run thread should update");
-	first.update_run_turn("run-a", "turn-a").expect("first run turn should update");
-	first.update_run_status("run-a", "succeeded").expect("first run status should update");
-
-	let state = first.inner.lock().expect("test should inspect the local cache");
-
-	assert!(
-		!state.events.contains_key("run-b"),
-		"run attempt updates should not refresh the full persistent event journal into the local cache"
-	);
-
-	drop(state);
-
-	let reopened = StateStore::open(&state_path).expect("reopened state store should open");
-	let attempt = reopened
-		.run_attempt("run-a")
-		.expect("run attempt lookup should succeed")
-		.expect("run attempt should persist");
-
-	assert_eq!(attempt.status(), "succeeded");
-	assert_eq!(attempt.thread_id(), Some("thread-a"));
-	assert_eq!(attempt.turn_id(), Some("turn-a"));
-	assert_eq!(reopened.event_count("run-b").expect("second event count should load"), 1);
-}
-
-#[test]
-fn persistent_project_run_listing_does_not_refresh_full_event_journal() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let observer = StateStore::open(&state_path).expect("observer state store should open");
-	let writer = StateStore::open(&state_path).expect("writer state store should open");
-
-	observer
-		.record_run_attempt("run-a", "PUB-101", 1, "running")
-		.expect("observer run should record");
-	observer
-		.upsert_lease("pubfi", "PUB-101", "run-a", IN_PROGRESS_STATE)
-		.expect("observer lease should record");
-	observer
-		.upsert_worktree("pubfi", "PUB-101", "x/pubfi-pub-101", "/tmp/worktrees/pub-101")
-		.expect("observer worktree should record");
-	observer.append_event("run-a", 1, "item/started", "{}").expect("observer event should append");
-	writer.record_run_attempt("run-b", "PUB-102", 1, "running").expect("writer run should record");
-	writer
-		.append_event("run-b", 1, "item/agentMessage/delta", "{}")
-		.expect("writer event should append");
-	writer
-		.record_run_attempt("run-c", "PUB-103", 1, "succeeded")
-		.expect("writer project run should record");
-	writer
-		.upsert_worktree("pubfi", "PUB-103", "x/pubfi-pub-103", "/tmp/worktrees/pub-103")
-		.expect("writer project worktree should persist");
-	writer
-		.append_event("run-c", 1, "thread/archive", "{}")
-		.expect("writer project event should append");
-
-	let mut writer_record = LinearExecutionEventRecord::new(
-		LinearExecutionEventIdentity {
-			service_id: "pubfi",
-			issue_id: "PUB-102",
-			issue_identifier: "PUB-102",
-			run_id: "run-b",
-			attempt_number: 1,
-		},
-		"closeout",
-		String::from("2026-04-29T10:12:00Z"),
-		"closeout",
-	);
-
-	writer_record.summary = Some(String::from("Writer closeout."));
-	writer_record.pr_url = Some(String::from("https://github.com/hack-ink/decodex/pull/102"));
-	writer_record.commit_sha = Some(String::from("2222222222222222222222222222222222222222"));
-
-	writer
-		.record_linear_execution_event(&writer_record)
-		.expect("writer ledger event should persist");
-
-	let runs = observer.list_leased_runs("pubfi").expect("leased runs should load");
-	let recent_runs = observer.list_recent_runs("pubfi", 10).expect("recent runs should load");
-	let leases = observer.list_active_shared_leases("pubfi").expect("shared leases should load");
-	let worktrees = observer.list_worktrees("pubfi").expect("worktrees should load");
-
-	assert_eq!(runs.len(), 1);
-	assert_eq!(runs[0].run_id(), "run-a");
-	assert_eq!(runs[0].event_count(), 1);
-	assert_eq!(runs[0].last_event_type(), Some("item/started"));
-	assert!(
-		recent_runs.iter().any(|run| run.run_id() == "run-c"
-			&& run.event_count() == 1
-			&& run.last_event_type() == Some("thread/archive")),
-		"project-scoped persistent event summaries should still load for matching runs"
-	);
-	assert_eq!(leases.len(), 1);
-	assert_eq!(leases[0].issue_id(), "PUB-101");
-	assert_eq!(worktrees.len(), 2);
-	assert_eq!(worktrees[0].issue_id(), "PUB-101");
-
-	let state = observer.inner.lock().expect("test should inspect the local cache");
-
-	assert!(
-		!state.events.contains_key("run-b"),
-		"operator run listing should refresh event summaries without materializing unrelated event rows"
-	);
-	assert!(
-		!state.events.contains_key("run-c"),
-		"operator run listing should refresh project summaries without materializing project event rows"
-	);
-	assert!(
-		!state.event_summaries.contains_key("run-b"),
-		"operator run listing should not refresh summaries for runs outside the requested project"
-	);
-	assert!(
-		!state.linear_execution_events.contains_key(&writer_record.idempotency_key),
-		"operator run and worktree listing should not refresh the full persistent ledger into the local cache"
-	);
-}
-
-#[test]
-fn persistent_project_listing_does_not_refresh_full_event_journal() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let observer = StateStore::open(&state_path).expect("observer state store should open");
-	let writer = StateStore::open(&state_path).expect("writer state store should open");
-	let registration = ProjectRegistration {
-		service_id: String::from("pubfi"),
-		config_path: temp_dir.path().join("project.toml"),
-		repo_root: temp_dir.path().join("repo"),
-		worktree_root: temp_dir.path().join("repo/.worktrees"),
-		workflow_path: temp_dir.path().join("repo/WORKFLOW.md"),
-		tracker_api_key_env_var: String::from("LINEAR_API_KEY_HACKINK"),
-		github_token_env_var: String::from("GITHUB_PAT_Y"),
-		enabled: true,
-		config_fingerprint: String::from("abc123"),
-		updated_at: String::from("2026-05-25T00:00:00Z"),
-		updated_at_unix: 1_779_667_200,
-	};
-
-	observer.upsert_project(&registration).expect("project should persist");
-	writer.record_run_attempt("run-b", "PUB-102", 1, "running").expect("writer run should record");
-	writer
-		.append_event("run-b", 1, "item/agentMessage/delta", "{}")
-		.expect("writer event should append");
-
-	let projects = observer.list_projects().expect("projects should load");
-
-	assert_eq!(projects.len(), 1);
-	assert_eq!(projects[0].service_id(), "pubfi");
-
-	let state = observer.inner.lock().expect("test should inspect the local cache");
-
-	assert!(
-		!state.events.contains_key("run-b"),
-		"project listing should not refresh the full persistent event journal into the local cache"
-	);
-	assert!(
-		!state.event_summaries.contains_key("run-b"),
-		"project listing should not refresh protocol summaries unrelated to the registry"
-	);
-}
-
-#[test]
-fn persistent_retry_budget_queries_do_not_refresh_full_event_journal() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let observer = StateStore::open(&state_path).expect("observer state store should open");
-	let writer = StateStore::open(&state_path).expect("writer state store should open");
-
-	writer
-		.record_run_attempt("run-a", "PUB-101", 1, "interrupted")
-		.expect("writer retry attempt should record");
-	writer.record_run_attempt("run-b", "PUB-102", 1, "running").expect("writer run should record");
-	writer
-		.append_event("run-b", 1, "item/agentMessage/delta", "{}")
-		.expect("writer event should append");
-
-	assert_eq!(observer.retry_budget_attempt_count("PUB-101").expect("retry count should read"), 1);
-	assert!(
-		observer
-			.issue_has_retry_budget_attempt_after("PUB-101", 0)
-			.expect("retry after query should read")
-	);
-	assert!(
-		!observer
-			.issue_has_retry_budget_attempt_after("PUB-101", 1)
-			.expect("retry after query should read")
-	);
-
-	let state = observer.inner.lock().expect("test should inspect the local cache");
-
-	assert!(
-		!state.events.contains_key("run-b"),
-		"retry-budget queries should not refresh the full persistent event journal into the local cache"
-	);
-	assert!(
-		!state.event_summaries.contains_key("run-b"),
-		"retry-budget queries should not refresh protocol summaries unrelated to the issue"
-	);
-	assert!(
-		!state.run_attempts.contains_key("run-a"),
-		"retry-budget queries should use issue-scoped persistent reads instead of a full runtime refresh"
-	);
-}
-
-#[test]
-fn persistent_shared_claim_check_does_not_refresh_full_event_journal() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let observer = StateStore::open(&state_path).expect("observer state store should open");
-	let holder = StateStore::open(&state_path).expect("holder state store should open");
-	let writer = StateStore::open(&state_path).expect("writer state store should open");
-	let slot_root = temp_dir.path().join("slots");
-
-	observer
-		.configure_dispatch_slot_root("pubfi", &slot_root)
-		.expect("observer slot root should configure");
-	holder
-		.configure_dispatch_slot_root("pubfi", &slot_root)
-		.expect("holder slot root should configure");
-	writer.record_run_attempt("run-b", "PUB-102", 1, "running").expect("writer run should record");
-	writer
-		.append_event("run-b", 1, "item/agentMessage/delta", "{}")
-		.expect("writer event should append");
-
-	assert!(
-		holder
-			.try_acquire_lease("pubfi", "PUB-101", "run-a", IN_PROGRESS_STATE)
-			.expect("holder should acquire the shared issue claim")
-	);
-	assert!(
-		observer
-			.issue_has_active_shared_claim("pubfi", "PUB-101")
-			.expect("shared claim check should read")
-	);
-
-	let state = observer.inner.lock().expect("test should inspect the local cache");
-
-	assert!(
-		!state.events.contains_key("run-b"),
-		"shared claim checks should not refresh the full persistent event journal into the local cache"
-	);
-	assert!(
-		!state.event_summaries.contains_key("run-b"),
-		"shared claim checks should not refresh protocol summaries unrelated to the issue"
-	);
-}
-
-#[test]
-fn persistent_linear_execution_event_listing_does_not_refresh_full_ledger() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let observer = StateStore::open(&state_path).expect("observer state store should open");
-	let writer = StateStore::open(&state_path).expect("writer state store should open");
-	let mut writer_record = LinearExecutionEventRecord::new(
-		LinearExecutionEventIdentity {
-			service_id: "pubfi",
-			issue_id: "PUB-102",
-			issue_identifier: "PUB-102",
-			run_id: "run-b",
-			attempt_number: 1,
-		},
-		"closeout",
-		String::from("2026-04-29T10:12:00Z"),
-		"closeout",
-	);
-
-	writer_record.summary = Some(String::from("Writer closeout."));
-	writer_record.pr_url = Some(String::from("https://github.com/hack-ink/decodex/pull/102"));
-	writer_record.commit_sha = Some(String::from("2222222222222222222222222222222222222222"));
-
-	writer
-		.record_linear_execution_event(&writer_record)
-		.expect("writer ledger event should persist");
-
-	let observed = observer
-		.list_linear_execution_events("pubfi", "PUB-102")
-		.expect("observer should read issue-scoped ledger events");
-
-	assert_eq!(observed, vec![writer_record.clone()]);
-
-	let state = observer.inner.lock().expect("test should inspect the local cache");
-
-	assert!(
-		!state.linear_execution_events.contains_key(&writer_record.idempotency_key),
-		"issue-scoped ledger listing should not refresh the full persistent ledger into the local cache"
-	);
-}
+mod cache_isolation;
+mod concurrency;
+mod protocol_replay;
+mod schema_migration;
